requirements.txt
streamlit
pandas
reportlab
openpyxl
python-dateutil
psycopg2-binary
